getStringWidth(): fixed unsigned char assumption
The function MiniGrafx::getStringWidth() assumed that the char type is
unsigned. It just takes a text[length] and calculates with it.

In environments where a char is signed, this can result in a negative
value, resulting of a segmentation fault because offset is negative in
MiniGrafx::readFontData(), causing pgm_read_byte() to read from
unassigned memory.

This patch adds an explicit cast to unsigned char, so the signedness of
chars does not matter anymore.